move Polyhedra imports to init.jl
