+++ src/AbstractPolytope.jl
@@ -186,40 +186,3 @@ end
 function _vertices_list(P::AbstractPolytope, backend)
     return vertices_list(P)
 end
-
-# =============================================
-# Functions that require Polyhedra
-# =============================================
-
-function load_polyhedra_abstractpolytope() # function to be loaded by Requires
-return quote
-import .Polyhedra: polyhedron
-export polyhedron
-using .Polyhedra: HRep, VRep,
-                  removehredundancy!, removevredundancy!,
-                  hreps, vreps,
-                  intersect,
-                  convexhull,
-                  hcartesianproduct, vcartesianproduct,
-                  points,
-                  default_library
-import JuMP, GLPK
-
-function default_polyhedra_backend(P, N::Type{<:AbstractFloat})
-    return default_library(LazySets.dim(P), Float64)
-end
-
-function default_polyhedra_backend(P, N::Type{<:Rational})
-    return default_library(LazySets.dim(P), Rational{Int})
-end
-
-function default_lp_solver(N::Type{<:AbstractFloat})
-    return JuMP.with_optimizer(GLPK.Optimizer)
-end
-
-function default_lp_solver(N::Type{<:Rational})
-    return JuMP.with_optimizer(GLPK.Optimizer, method=:Exact)
-end
-
-end # quote
-end # function load_polyhedra_hpolytope()
